+++ packages/libredwg-converter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mlightcad/libredwg-converter
 
+## 2.0.7
+
+### Patch Changes
+
+- feat: regularize block space name and paper space name
+- Updated dependencies
+  - @mlightcad/data-model@1.1.7
+
 ## 2.0.6
 
 ### Patch Changes